agent/requirements.txt
certifi==2018.8.24
chardet==3.0.4
gevent==1.3.6
greenlet==0.4.14
idna==2.7
ifaddr==0.1.4
ipaddress==1.0.22
netifaces==0.10.7
scapy==2.4.0
websockets==6.0
python-nmap==0.6.1
requests==2.21.0
python-iptables==0.14.0

